pyproject.toml
[tool.poetry]
name = "ai-planet-backend"
version = "0.1.0"
description = "backend service to process pdfs with nlp"
authors = ["codeblech <my.malikyash@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.6"}
fastapi-limiter = "^0.1.6"
aiofiles = "^24.1.0"
pytest = "^8.3.4"
pytest-asyncio = "^0.25.0"
sqlalchemy = "^2.0.36"
python-dotenv = "^1.0.1"
google-generativeai = "^0.8.3"
chromadb = "^0.5.23"
pypdf = "^5.1.0"


[tool.poetry.group.dev.dependencies]
sphinx = "^8.1.3"
furo = "^2024.8.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
